HooksTest: split `register()` specific tests to dedicated test class
